backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis and Celery
redis==5.0.1
celery==5.3.4

# AI/ML
openai==1.10.0
httpx==0.26.0
pillow==10.2.0
numpy==1.26.3

# Vector DB
pinecone-client==3.0.0
qdrant-client==1.7.0

# Storage
boto3==1.34.14
python-magic==0.4.27

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1

# PDF Generation
reportlab==4.0.8
weasyprint==60.2

